# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.Chain:
    properties:
      coin_gecko_id:
        type: string
      currency:
        type: string
      id:
        type: integer
      name:
        type: string
      short_name:
        type: string
    type: object
...
  model.GuildCustomCommand:
    properties:
      actions:
        items:
          type: integer
        type: array
      channels_permissions:
        items:
          type: integer
        type: array
      cooldown:
        type: integer
      cooldown_duration:
        type: integer
      description:
        type: string
      enabled:
        type: boolean
      guild_id:
        type: string
      id:
        type: string
      roles_permissions:
        items:
          type: integer
        type: array
    type: object
...
  model.Token:
    properties:
      address:
        type: string
      chain:
        $ref: '#/definitions/model.Chain'
      chain_id:
        type: integer
      coin_gecko_id:
        type: string
      decimal:
        type: integer
      discord_bot_supported:
        type: boolean
      guild_default:
        type: boolean
      id:
        type: integer
      is_native:
        type: boolean
      name:
        type: string
      symbol:
        type: string
    type: object
...
  request.GiftXPRequest:
    properties:
      admin_discord_id:
        type: string
      channel_id:
        type: string
      guild_id:
        type: string
      user_discord_id:
        type: string
      xp_amount:
        type: integer
    type: object
...
  request.UpsertCustomTokenConfigRequest:
    properties:
      active:
        type: boolean
      address:
        type: string
      chain:
        type: string
      chain_id:
        type: integer
      coin_gecko_id:
        type: string
      decimals:
        type: integer
      discord_bot_supported:
        type: boolean
      guild_default:
        type: boolean
      guild_id:
        type: string
      id:
        type: integer
      name:
        type: string
      symbol:
        type: string
    type: object
...
  response.CreateCustomCommandResponse:
    properties:
      data:
        $ref: '#/definitions/model.GuildCustomCommand'
    type: object
...
  response.GetCustomCommandResponse:
    properties:
      data:
        $ref: '#/definitions/model.GuildCustomCommand'
    type: object
...
  response.GiftXpHandlerResponse:
    properties:
      data:
        $ref: '#/definitions/response.HandleUserActivityResponse'
    type: object
...
  response.HandleUserActivityResponse:
    properties:
      action:
        type: string
      added_xp:
        type: integer
      channel_id:
        type: string
      current_level:
        type: integer
      current_xp:
        type: integer
      guild_id:
        type: string
      level_up:
        type: boolean
      timestamp:
        type: string
      user_id:
        type: string
    type: object
...
  response.ListAllCustomTokenResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Token'
        type: array
    type: object
...
  response.ListCustomCommandsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.GuildCustomCommand'
        type: array
    type: object
...
  response.ResponseMessage:
    properties:
      message:
        type: string
    type: object
...
  response.UpdateCustomCommandResponse:
    properties:
      data:
        $ref: '#/definitions/model.GuildCustomCommand'
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a swagger for mochi api.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger API
  version: "1.0"
paths:
...
  /configs/custom-tokens:
    post:
      consumes:
      - application/json
      description: Guild custom token config
      parameters:
      - description: Custom guild custom token config request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.UpsertCustomTokenConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseMessage'
      summary: Guild custom token config
      tags:
      - Config
...
  /gift/xp:
    post:
      consumes:
      - application/json
      description: Gift xp handler
      parameters:
      - description: Gift XP handler request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.GiftXPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GiftXpHandlerResponse'
      summary: Gift xp handler
      tags:
      - Gift
...
  /guilds/{guild_id}/custom-commands:
    get:
      consumes:
      - application/json
      description: List custom commands
      parameters:
      - description: Guild ID
        in: path
        name: guild_id
        required: true
        type: string
      - description: Enabled
        in: query
        name: enabled
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ListCustomCommandsResponse'
      summary: List custom commands
      tags:
      - Custom Command
    post:
      consumes:
      - application/json
      description: Create custom command
      parameters:
      - description: Guild ID
        in: path
        name: guild_id
        required: true
        type: string
      - description: Create custom command request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/model.GuildCustomCommand'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CreateCustomCommandResponse'
      summary: Create custom command
      tags:
      - Custom Command
  /guilds/{guild_id}/custom-commands/{command_id}:
    delete:
      consumes:
      - application/json
      description: Delete custom commands
      parameters:
      - description: Guild ID
        in: path
        name: guild_id
        required: true
        type: string
      - description: Command ID
        in: path
        name: command_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Delete custom commands
      tags:
      - Custom Command
    get:
      consumes:
      - application/json
      description: Get custom commands
      parameters:
      - description: Guild ID
        in: path
        name: guild_id
        required: true
        type: string
      - description: Command ID
        in: path
        name: command_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GetCustomCommandResponse'
      summary: Get custom commands
      tags:
      - Custom Command
    put:
      consumes:
      - application/json
      description: Update custom command
      parameters:
      - description: Guild ID
        in: path
        name: guild_id
        required: true
        type: string
      - description: Command ID
        in: path
        name: command_id
        required: true
        type: string
      - description: Update custom command request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/model.GuildCustomCommand'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UpdateCustomCommandResponse'
      summary: Update custom command
      tags:
      - Custom Command
...
  /guilds/custom-tokens:
    get:
      consumes:
      - application/json
      description: List all guild custom token
      parameters:
      - description: Guild ID
        in: path
        name: guild_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ListAllCustomTokenResponse'
      summary: List all guild custom token
      tags:
      - Guild
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: access_token
    type: apiKey
swagger: "2.0"