# pkg/response/response.go
package response

type ResponseMessage struct {
	Message string `json:"message"`
}

type ResponseSucess struct {
	Success bool `json:"success"`
}

type ResponseStatus struct {
	Status string `json:"Status"`
}